+++ cellxgene_schema_cli/tests/test_remove_labels.py
@@ -1,6 +1,6 @@
 import unittest
 
-from cellxgene_schema.remove_labels import remove_labels_inplace
+from cellxgene_schema.remove_labels import AnnDataLabelRemover
 from fixtures.examples_validate import (
     adata_with_labels,
     adata,
@@ -11,21 +11,22 @@ from pandas.testing import assert_frame_equal
 class TestRemoveLabels(unittest.TestCase):
 
     def setUp(self):
-        self.adata_with_labels = adata_with_labels.copy()
+        self.anndata_label_remover = AnnDataLabelRemover(adata_with_labels.copy())
         self.adata_no_labels = adata
 
-    def test_remove_labels_inplace(self):
-        self.adata_with_labels.raw = adata_with_labels
-        self.adata_with_labels.raw.var.drop("feature_is_filtered", axis=1, inplace=True)
-        remove_labels_inplace(self.adata_with_labels)
+    def test_remove_labels(self):
+        self.anndata_label_remover.adata.raw = adata_with_labels
+        self.anndata_label_remover.adata.raw.var.drop("feature_is_filtered", axis=1, inplace=True)
+        self.anndata_label_remover.remove_labels()
+        adata_labels_removed = self.anndata_label_remover.adata
+        assert_frame_equal(adata_labels_removed.obs, self.adata_no_labels.obs)
+        assert_frame_equal(adata_labels_removed.var, self.adata_no_labels.var)
+        assert_frame_equal(adata_labels_removed.raw.var, self.adata_no_labels.raw.var)
 
-        assert_frame_equal(self.adata_with_labels.obs, self.adata_no_labels.obs)
-        assert_frame_equal(self.adata_with_labels.var, self.adata_no_labels.var)
-        assert_frame_equal(self.adata_with_labels.raw.var, self.adata_no_labels.raw.var)
+    def test_remove_labels_no_raw(self):
+        self.anndata_label_remover.remove_labels()
+        adata_labels_removed = self.anndata_label_remover.adata
 
-    def test_remove_labels_inplace_no_raw(self):
-        remove_labels_inplace(self.adata_with_labels)
-
-        self.assertIsNone(self.adata_with_labels.raw)
-        assert_frame_equal(self.adata_with_labels.obs, self.adata_no_labels.obs)
-        assert_frame_equal(self.adata_with_labels.var, self.adata_no_labels.var)
+        self.assertIsNone(adata_labels_removed.raw)
+        assert_frame_equal(adata_labels_removed.obs, self.adata_no_labels.obs)
+        assert_frame_equal(adata_labels_removed.var, self.adata_no_labels.var)